third_party/WebKit/LayoutTests/editing/deleting/delete-character-001.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable>fo|o   bar    baz</div>',
        'delete',
        '<div contenteditable>f|o   bar    baz</div>'),
        'Delete a character');
</script>
